feature: support SetLoopback()
* create a separate file for errors (Metrics/ClassLength)
